+++ src/payment/WalletPayment.java
@@ -1,10 +1,8 @@
 package payment;
 
-import payment.PaymentMethod;
-
 public class WalletPayment implements PaymentMethod {
     @Override
-    public void pay(int amount) {
+    public void pay(float amount) {
         System.out.println("wallet payment");
     }
 }
